requirements.txt
django==4.2.7
transformers==4.35.2
torch==2.1.0
pyannote-audio==3.1.0
openai-whisper==20231117
pydub==0.25.1
ffmpeg-python==0.2.0
numpy==1.26.4  # Add this to pin NumPy <2.0 for compatibility
keybert==0.8.5
requests==2.32.3